Bug fixed.

+++ TicTacToeREST/src/main/java/be/ltmartin/tictactoerest/Constants.java
@@ -5,7 +5,7 @@ public interface Constants {
     byte O_WINS = 0;
     byte TIE = 2;
     byte CONTINUE = 3;
-    byte INITIAL_COORDINATE = Byte.MAX_VALUE;
+    Integer INITIAL_COORDINATE = Integer.MAX_VALUE;
     Character EMPTY_CHARACTER = 'E';
     Character X_MARK = 'X';
     Character O_MARK = 'O';
